softmmu/physmem: Introduce physical guest memory debug interface
The new interface MemoryDebugOps which hook into guest
physical and virtual memory debug interfaces to allow vendor
specific assist/hooks for debugging and delegating accessing
the guest memory. This is required for example in case of
AMD SEV platform where the guest memory is encrypted and a
SEV specific debug assist/hook will be required to access
the guest memory.

The MemoryDebugOps are default to address_space_read and
address_space_write_rom for work with exist legacy VM.

Yuan Yao:
  - Exports the physical_memory_debug_ops variable for functions
    in target/i386/helper.c
  - Refactor the too long subject line
  - Rewrite the commit message

Signed-off-by: Ashish Kalra <ashish.kalra@amd.com>
Signed-off-by: Yuan Yao <yuan.yao@intel.com>