i386/tdx: define tdx_fixed0/tdx_fixed1 for each feature word
When Intel TDX module virtualizes CPUID to guest TDs, One kind of
virtualization type is Fixed. TD guest would see the fixed CPUID values
(0 or 1) according to the chapter 16.2 in [1]. Some of the feature bits
are not supported or published yet, add a placeholder bitmask to
represent them.

Besides, some PV feature bits are not supported in TD guest, see them as
fixed0 type.

[1] TDX Module spec
https://software.intel.com/content/dam/develop/external/us/en/documents/tdx-module-1eas-v0.85.039.pdf

Signed-off-by: Chenyi Qiang <chenyi.qiang@intel.com>