kvm/tdx: Locate TDX metadata via a GUID-ed structure
TDX metadata is located by a TDX metadata offset block, which
is a GUID-ed structure. The data portion of the GUID structure
contains only an 4-byte field that is the offset of TDX metadata
to the end of firmware file.

Select X86_FW_OVMF when TDX is enable to leverage existing functions
to parse and search OVMF's GUID-ed structures.

Note that the firmware file is mmap-ed temporarily as
pc_system_parse_ovmf_flash() accepts a pointer and a size.

Signed-off-by: Chao Gao <chao.gao@intel.com>