docs/interop/firmware.json: Add TDX support
Qemu use generic loader device to load firmware for TD-guest which
is different from normal and SEV guest.

Create enum definition '@loader' and '@intel-tdx' for TDX and add
documention for the new enums. Add an example that shows some of
the requirements for TDX, including not having SMM support and the
requirement for an X64-only build.

Signed-off-by: Zhenzhong Duan <zhenzhong.duan@intel.com>
Signed-off-by: Isaku Yamahata <isaku.yamahata@intel.com>