kvm/tdx: Add properties for TD attributes
In TDX specification, there is a 64bit parameter called TD attributes to
specify its characteristic.  For example, debug (bit 0).  If the bit is
set, the guest TD can be debugged.  (Guest state isn't protected.  It
shouldn't be set for production use.)  Part of these bits is reserved for
future use and more attributes will be added in the future.

Add sept-ve-disable, pks and kl attributes which are defined in the TDX
specification.  Because Linux TDX guest requires sept-ve-disable by default,
the bit is default on.  Other two bits are default off.

example:
qemu-system-x86_64 \
-object tdx-guest,id=tdx0,debug=on,sept-ve-disable=on,pks=off,kl=off

Signed-off-by: Isaku Yamahata <isaku.yamahata@intel.com>