kvm/tdx: Support sept_ve_disable and pks for TD attributes
In TDX specification, there is a 64bit parameter called TD attributes to
specify its characteristic.  For example, debug (bit 0).  If the bit is
set, the guest TD can be debugged.  (Guest state isn't protected.  It
shouldn't be set for production use.)  Part of these bits is reserved for
future use and more attributes will be added in the future.

Support sept-ve-disable and pks attributes which are defined in the TDX
specification.  Add sept-ve-disable property.  For PKS, cpu support is
specified as cpu feature.  Set PKS TD attributes bit if PKS feature is
enabled by cpu feature.

example:
qemu-system-x86_64 -cpu pks=on \
-object tdx-guest,id=tdx0,sept-ve-disable=on

Signed-off-by: Isaku Yamahata <isaku.yamahata@intel.com>