[REVERTME] kvm/tdx: Make GetQuote handler accept old ABI
TDG.VP.VMCALL<GetQuote> in GHCI1.0 spec is changed so that the shared GPA
size is passed in R13.  With The old GHCI R13 isn't used and the buffer size
is assumed to be fixed 8KBytes.

If R13 input value is zero, it assumes the old GetQuote ABI and handle it
for compatibility.

Signed-off-by: Isaku Yamahata <isaku.yamahata@intel.com>